+++ src/components/claims/claim-details/components/ClaimNotes.tsx
@@ -9,10 +9,9 @@ import { Spinner } from 'hedvig-ui/sipnner'
 import React, { useState } from 'react'
 
 import styled from '@emotion/styled'
-import { Button, CardContent } from '@hedvig-ui'
+import { Button, CardContent, List, ListItem } from '@hedvig-ui'
 import { PaperTitle } from 'components/claims/claim-details/components/claim-items/PaperTitle'
 import { FadeIn } from 'hedvig-ui/animations/fade-in'
-import { List, ListItem } from 'hedvig-ui/list'
 import { Spacing } from 'hedvig-ui/spacing'
 import { TextArea } from 'hedvig-ui/text-area'
 import { Paragraph, Shadowed } from 'hedvig-ui/typography'
